KS-109 reorganize class Request. Moved private methods.
